cache: add cache for files and directory sizes
This helps not needed to read directories over and over again or having to compute directory sizes several times.
The is initialized as a struct. Once asked a value, it checks if the value is computed and returns it if it is, otherwise
computes the value first and then returns it.

Had to abolish some parallelism in top_items for that but I might be able to paralleize some cache operations in the future.

Closes #16