local: fix crash when calling "local" on a target dir being modified by cargo
I was collecting files, filtering out nonexisting paths and then (in parallel) requesting metadata on
the files but sometimes it happened that temporary files were deleted by cargo between the collection of
available paths and a rayon job asking for the metadata leading fs::metadata() to fail unwrapping as the
file was already gone.
Mitigate by only asking for metadata if the file still exists right before doing so.

Fixes #43